mnemonic-rag/pyproject.toml
[tool.poetry]
name = "mnemonic-rag"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [
    { include = "app" },
]

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
uvicorn = "^0.23.2"
langserve = {extras = ["server"], version = ">=0.0.30"}
pydantic = "<2"
retrieval-agent = {path = "packages/retrieval-agent", develop = true}
self-query-qdrant = {path = "packages/self-query-qdrant", develop = true}


[tool.poetry.group.dev.dependencies]
langchain-cli = ">=0.0.15"
qdrant-client = "==1.7.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
